requirements.txt
EbookLib==0.20
pillow==12.0.0
PyMuPDF==1.26.7
weasyprint==67.0
